test(integration/ws-healthcheck): refactoring to use testcontainers
